rename: data 및 constants 파일명 카멜 케이스 -> 케밥 케이스
